1. Added itoa/atoi functions. 2. Moved string fuctions.
