+++ examples/full_example.c
@@ -99,7 +99,7 @@ int main()
     
     // parsing test
     emJSON_clear(&test);
-    char str_input[] = "{\"sensor1\":0.045600E3,\"message\":\"JSON Is Cool\",\"sensor2\":142}";
+    char str_input[] = "{\"sensor1\":0.045600E3,\"message\":\"JSON Is Cool\",\"sensor2\":-1423}";
     json_parse(&test, str_input);
     // Serialize
     printf("==JSON Serialize Test==\n");
